AddLinkGatePassageCompetitioner

+++ Models/Sector.cs
@@ -9,9 +9,9 @@ namespace CompetitionResults.Models
     public class Sector : BaseModel
     {
         public int Number { get; set; }
-        public long TrackId { get; set; }
+        public int TrackId { get; set; }
         public bool IsFull { get; set; }
         public virtual Track Track { get; set; }
-        public virtual List<GateWithPenalty>GatesWithPenalty { get; set; }
+        public virtual List<GateWithPenalty> GatesWithPenalty { get; set; }
     }
 }
